[WOOR-224] refactor: Facade 패턴 적용
